separated class navigation_items
